Denormalize Icla Signatures
After careful consideration and conversations, the ICLA Signature
record should hold the personal information supplied, not the user or
an association of the User model. This is to protect the integrity of
the data and to keep accurate records separate from the user.

In the future, the user will be able to modify her profile
information, but we need to retain the original information submitted
with the ICLA. A user should be permitted to update her personal
information, but only administrators should be permitted to update the
ICLA information. For this reason, this commit denormalizes the
information from the user object and the Icla Signature object. The
Icla Signature form is pre-populated with information that exists on
the user object, but that data is not tied to the user explicitly.

Additionally, aside from the legal requirements of the ICLA,
collecting and storing address information on the user did not seem
like a user case worth supporting. The address model and tables have
been removed and those fields are now properties of the Icla Signature
model.