Add basic authorization
This commit adds the base structure for using policy objects with
the authorization model. Individual authorization policies live in
+app/policies/authorizers+ and have a 1-1 relationship with models.

Thank you to this blog post: http://www.elabs.se/blog/52, which
provided the inspiration for this pattern.

NOTE: Some of the tests are currently failing because of the new
authentication model. They will be fixed in a future commit.