Associate new OmniAuth logins with existing users
The +.from_oauth+ method now accepts an optional second parameter, a User
object, to associate with the new account. If passed, the new Account
record will belong to the given user. If the parameter is omitted (or +nil+),
a new User is created too.

Accounts search Email records to see if an email address corresponds to
information contained in the OmniAuth hash. If an Email exists, the new
Account object will be associated with the User object associated with the
Email record.

Accounts search IclaSignature records to see if an email address corresponds
to information contained in the OmniAuth hash. If an Icla Signature exists
with the given email address, the new Account object will be associated with
the User object associated with the Icla Signature record.