Convert CanCan-style view helper to Pundit-style.
Document the Supermarket::Authorization module to explain how to
interface with Pundit.

The Pundit-style view helper for authorization is to pass the resource
being authorized into the policy method and call the action predicate.

For example to check if the current_user can create a new user, you
would do:

  policy(user).create?