Add Gravatar support for Users.
Creates UserHelper#gravatar_for which takes a User as a parameter and an
optional options hash. The options hash may contain the size for the gravatar,
which defaults to 48.

The method determines the Gravatar based on the user's email.

Displays the user's Gravatar on their profile.

Example use in a view:

  <%= gravatar_for current_user, size: 72 %>

or

  <%= gravatar_for current_user %>

Pivotal Tracker ID: 63206080
https://www.pivotaltracker.com/story/show/63206080