Only allow user to have single email address
This removes the current model where users can have
many email addresses to only a single email address defined
on the user itself. This may change in the future but currently
there's no mechanism for the user to manage multiple email addresses
so there's no reason to support this in the data model.