Feature specs check selectors instead of content
Checking content is a cheap way to gain confidence that the view is
working correctly, but as the previous commit shows, content assertions
are brittle. If it is truly important to fail the build when content
changes, we should be writing view specs. The basic user-facing
indication of success is the successful flash message, and that's what
we now test instead.