Enable Travis CI Slack Notifications
This enables Travis CI to notify Slack the token is semi-secret
so there's no real need to encrypt it. As stated in the Slack docs:

"Your integration token is semi-secret. The worst someone could do
if they had it was post Travis payloads to your channel. They could
not read any data."