Add sidekiq recipe
Like the Redis recipe, the Sidekiq recipe is minimally configurable;
just enough to get something running for now.

When deploys start, we tell Sidekiq workers to stop accepting new jobs.
When we stop our web server, we also stop Sidekiq. When we start the web
server, we start Sidekiq.

At some point, we'll want a process manager to handle some things that
Sidekiq currently does on its own: daemonization, and log redirection in
particular. It'd also be nice to have a monitor to revive crashed
processes.