Add link GitHub account action
Currently users are just dumped to /users/:id when they are prompted to
link their GitHub account and then just have to figure out how to. This
adds a new action at /profile/link-github that just provides a nice big button
that starts the OAuth process. Users are now dumped to this action if they attempt
to sign a CLA without having a linked GitHub account.