CCLA table expanding fix
This fixes an issue where when you click view contributors on the CCLA list
it would jump to the top of the page as it's an anchor link. This just
adds preventDefault to tame this behavior. It also changes the binding class
for the click event so it doesn't conflict with the Foundation expand class.