Implement the getting a cookbook version via API.
Create an API endpoint to hit to get a specific cookbook version. If that cookbook
or cookbook version are not found, a 404 is returned.

Create the CookbookVersion model and expand on getting a Cookbook version
through the API. A CookbookVersion is an artifactual record of a specific
version of a Cookbook.