Cookbook detail pages display the latest README
Supermarket extracts the README the same way it extracts metadata.json.
If there is no file in the tarball whose downcased name includes
'readme', it bails on the sharing request.

If the README's extension indicates the README is formatted as Markdown,
Supermarket converts that Markdown into HTML. Otherwise it assumes the
text is pre-formatted.