Users can download cookbooks
Supermarket exposes a consistent URL scheme for a given cookbook, where
each download URL redirects the user to the URL of the cookbook
artifact. Supermarket tracks the total number of downloads both in the
aggregate and per-version.

Note that when a user downloads a cookbook without explicitly specifying
a version, we assume this user would like to download the latest
version.