Show supported platforms on the cookbook detail page
Supported platforms are contained in the metadata.json file. Note that
the CookbookUpload::Metadata is now strict in its coercion. This is so
that Supermarket can catch and report errors with the metadata.json
"platforms" value format.

For now, SupportedPlatform records are directly associated to
CookbookVersion records. In the future, particularly when we want to
search and/or filter on the SupportedPlatform name attribute, we may end
up normalizing this records so that the relationship is indirect.