Implement cookbook following/unfollowing
Implement the ability for users to follow and unfollow a cookbook.
Signed in users may click Follow and they are associated to the cookbook
via CookbookFollower clicking Unfollow will delete that association.
When signed out the follow button is still displayed but the user is
directed to sign in then redirected back to the cookbook that they
innitially tried to follow.