Implement dropdown versions
Rather than displaying the cookbook versions in the sidebar display them
as a dropdown in the cookbook header. It's less important to see a list
of versions in the sidebar than it is to be able to easily switch
between them. The dropdown in the header will allow the user to easily
see what version they're currently viewing and be able to quickly
switch between versions. This also has a nice side effect of greatly
reducing the complexity of the JS, removing it completely or rather
leaning on Foundation.