Add cookbook follower notifications
Adds cookbook follower notifications and allows users to opt out of
notificaitons, adds segment.io events when someone follows/unfollows a cookbook
and reworks the mailers a bit to use a global layout and Sidekiq delayed
extensions. Also splits of ICLA and CCLA signature notification emails
into separate methods in the ClaSignatureMailer as it's a bit cleaner and will
allow for the emails to be different if need be.