Create the dashboard & welcome views.
The dashboard is what an authenticated user sees. It displays the activity feed
for the cookbooks they follow and easy to find links for the cookbooks they own
and contribute to.

The welcome page is what a non-authenticated user sees. It introduces them to
what Supermarket is and highlights what they can do on Supermarket.

The content is not all fleshed out yet, but it will soon include things like
featured cookbooks, cookbook activity feeds and the cookbooks that a user owns
and contributes to.