Overhaul and update the UI.
- Remove the cookbook categories nav in the cookbooks directory. The category
  nav is of questionable value, so for the time being the nav is being removed.
- Use the actual cookbook category on the cookbook partial.
- Make the search bar global with new styles.
- Remove the temp stats from the cookbooks directory as they now live on the
  welcome page.
- Update the layout of the cookbook partial
- Update the global header styles
- Add support for the generic platform for cookbooks.

Notes:

Wrap submit_form feature helpers in within blocks to be more specific about the
submit button be pressed. This needs to happen because the search bar form as
been added, meaning there is always at least one form on every page.

The names displayed on the cookbook partial are generated from Faker. They need
to be properly displayed based upon the cookbook's owner once that is
implemented.