Don't 500 when CCLA signature is invalid
For some reason it was assumed that CclaSignature#sign! returned
a boolean if the transaction failed, this was incorrect. It raises an
invalid record exception so we must rescue this in the controller then
render 'new' with any validation errors. This also adds abide validation
to the company field as that's what originally caused the bug.