Don't send cookbook notifications to imported users
Since we'll be running Supermarket in parallel with the current
community site we don't want to send cookbook follower notifications
from Supermarket until the user signs in, at least during the soft launch
period. Imported users will have a oauth_token for their chef_oauth2
account set to imported until they sign in so we can use that as an
indicator for whether or not they are active on Supermarket.