Don't use abide alpha_numeric for CLA fields
The abide alpha_numeric pattern doesn't allow spaces which is undesired
behavior for the CLA fields. This didn't seem to be the case in previous
versions of Foundation but it does seem to be the desired behavior
(there was a bug filed surrounding this issue). We can add stricter
pattern matching in the future but just ensuring that required
fields are filled out feels like enough.