Add the Supermarket Beta banner.
The Supermarket Beta banner is a global drop down that contains information
surrounding the Supermarket Beta. This banner will only be around until the Beta
ends.

The content contains two key pieces that are displayed side-by-side:

- what users should expect
- how users can get involved

The banner features a slide toggle when pressed and emulates the look and feel
of buttons throughout the application. On mobile, the two pieces of content
collapse into one stacked column.