Create a worker which refreshes OC-ID OAuth tokens
In order to refresh OC-ID OAuth tokens, we need to store the refresh
token provided by OC-ID. We also need to store the timestamp provided by
OC-ID which represents the time at which the given OAuth token will
expire. This commit adds a column to the accounts table to store the
former, and adds the code required to store both whenever a user
authenticates.