[WIP] use sprockets for stylesheets
instead of importing all of the styles into one stylesheet, let's use sprockets
so that we do not have to recompile every library and stylesheet when there is a
change. let's only recompile what we need to.

this also begins to break up the larger stylesheets into multiple files while
re-organizing where they are located.

this is not done yet, but it is getting pretty close

Only import what we need to and transition foundation settings into styles.

this is still wip, and I realize there are now tabs in the mix. will fix and
remove when I finish cleaning this up.