Move styles from SCSS import to Sprockets manifest
Importing all of the third-party libraries and Supermarket stylesheets into one
SCSS file means that whenever a single file is changed, every single dependency
and stylesheet in Supermarket needs to be compiled as well. This results in slow
development cycle when working on styles.

This commit is the first small step to improving the way our styles are loaded
and compiled into the app - it moves the styles from a global import to a
Sprockets manifest file. This means that in development, there are a handful of
seperate stylesheets. In production, they get compiled and compressed. There is
a common.scss file that handles the importing of needed depedencies for each
file. The common file includes parts of Foundation and most of compass, as well
as some variables and mixins.

Supermarket's reliance upon importing many parts of Foundation ties the
application very tightly to Foundation. It would be ideal to move away from
Foundation and override its styles within Supermarket instead of changing its
defaults.

In the future, the goal is to only import what is needed into each Supermarket
stylesheet. This will improve the compile time for individual files, as well as
make it easier to debug styles based on what stylesheet they live in while using
development.

Another goal is to split up the larger stylesheets into directories with
multiple files, making the stylesheets more digestible and easier to work with.

This commit, however, does not accomplish those goals. It simply begins the
transition to allow those goals to be possible.