Remove rake spec:javascipts from Travis CI config.
At this point in time, Travis CI takes about 15 minutes to run the entire test
suite. Travis CI is running the JavaScript specs, which requires installing a
bunch of NPM packages. The thing is we only have one JavaScript spec for the
alert message to disappear. Removing the rake spec:javascripts script from
Travis CI runs should improve the time it takes to complete. While this is not
ideal once/if more JS is added to the app, I think for now it is a little win
that helps.