Render follow/unfollow button via AJAX.
Currently the follow, unfollow actions redirect back. This causes a bug
in Chrome where items in the history stack are duplicated causing the
user to have to click back twice. This changes the follow, unfollow
button to be a remote link with a ajax:success event bound to it that
loads the current view and loads the follow, unfollow button from that
into the dom without a page reload. This doesn't really save on page
load as it's still rendering the entire view but it does fix the bug and
provide a better user experience.