Only import SCSS required for each file, remove settings.scss
To try and improve the performance of our styles compilation this
changes the application style sheets to only import the SCSS that is
required on a per file basis. This also removes the reliance on
settings.scss in favor of overriding foundation.css styles with
regular CSS overrides so we don't need to import as much.