Improve/expand Supermarket migration metrics
The previous metrics gave a helpful but simplistic overview of the
status of the import by comparing raw Supermarket counts to raw
Community Site counts. The new metrics show how many records of various
types were imported into Supermarket and roughly how many records could
potentially be imported from the Community Site.

In the case of cookbook versions and consequently supported platforms,
the comparison will not be 1-1. This is because there are a dozen or so
cookbooks on the community site which have duplicate cookbook versions
with the same version number. These duplicate cookbook versions also
have duplicate associated platform versions records. Because we don't
import the duplicate records, the numbers don't match up exactly.