Move the CLA companies contributors listing into their own view.
Currently, the way contributors on behalf of a company are displayed is a
show/hide toggle button with in the table of companies. This results in a funky
looking and difficult to read table. This changes the button to display the
contributors to link to a new view that lists all of the contributors.

Above that table is a button to return to the companies listing, as well as
context sensitive buttons:

- one for viewing the CCLA if the current user is a member of that company
- one for going to the invitations management view for that company