Create partial for CCLA org nav tabs.
In order to reuse the CCLA org nav tabs for viewing the CCLA signature, viewing
Contributors and managing Contributors, move it all into a partial that can be
used where we need it. This partial includes logic with authorizers for
determining who can see what.

One place this is being used in a different way is when a user is viewing the
list of Contributors for an Organization. If the user is part of that
Organization, they will see the nav tabs for that Organization. If the user is
not part of that Organization, they will see a link bringing them back to the
list of Organizations (companies).