Remove CLA signature links from user nav.
To keep the user nav as simple possible, there should only be absolutely
necessary links in the nav. I believe that the ICLA and CCLA nav items are not
appropriate to be present. For example, a user could use Supermarket without
ever needing to sign a CLA.

The flow for signing a CLA, imo should be: go to Contribute where you can view
other contributors, info about contributing and elect to become a contributor as
an individual or on behalf of a company. If a person opens a PR on a Chef repo,
Curry will let them know that they need to sign the CLA. These seem like the two
most common and useful ways for people to discover CLA signing without it being
all up in their user nav.