Create the Become a Contributor page
This re-organizes the file structure of where the contribute/sidebar is located,
while creating the route, controller action and spec for the page to point
people to in order to understand what it means to become a contributor.

The content is not ready in this commit, but the framework is there.