Allow mailer URLs to have protocol specified
You cannot rely on email clients handling URLs without a protocol
specified correctly. So this allows a protocol to be specified
(generally either http or https) so a full absolute URL can be
generated. This issue originally reared its head by not routing to
images correctly in most email clients.