Combine OrganizationAuthorizer and CclaSignatureAuthorizer
Since the organization is the gatekeeper of CCLA signatures it seems
logical to move the the authorization logic to OrganizationAuthorizer
this also allows the logic to be shared for organizationy type actions
like managing contributors. This also adds a new authorized action,
resign_ccla? which protects a CCLA from being resigned by a regular
contributor in an organization.