Change updated date back to Cookbook#updated_at
Turns out that a cookbook is saved when a new version is published so we
should just use Cookbook#updated_at to determine when a cookbook was
most recently updated. Using the CookbookVersion#created_at date seems
good in theory but in practice it makes it a pain to order cookbooks by
most recently updated since then you need to get subqueries involved.