Display organization members as contributors.
The implementation of User.all_cla_signers did not account for the fact
that users who are contributors on behalf of an organization are
authorized to contribute to Chef projects. This changes that method to
be User.authorized_contributors, which is all of the users who have
signed the ICLA or the CCLA or are a contributor on behalf of an
organization.