universe should return opscode location type
Chef plans to redirect the cookbooks/community site URIs to
supermarket in order to facilitate the launch. Currently, returning
"supermarket" as the location type means that berkshelf (the client)
cannot download cookbooks from supermarket.

This change sets the location type to opscode by defining a new
constant, which can be updated at at later point if/when necessary.