Fix specs broken by #eda127b
We don't need the spec that checks for route compliance for this
particular method since we're not generating a full route.

Convert the port to a string before checking if it's empty. We could
use #blank? here, but @jtimberman avoids using ActiveSupport in order
to not develop habits that might leak into cookbook code.