Expect the proper string for berks
Berkshelf's `CommunityREST` expects a particular string, so we use
`#eql` to check that `download_path` generates the right thing with
various options. We also want to make sure if a port and protocol
isn't specified, we get the right thing (`http` by default, and no `:`
w/ an empty port string).