Have Cookbook#description return nil
There is an edge case where latest_cookbook_version can be nil because
it was deleted after a Cookbook has been initialized so if description
is called it's delegated to latest_cookbook_version which is nil and it
raises an exception. This enables the allow_nil option to the
description delegate so rather than raising an exception description
will just return nil.