Add more useful account error messages
Currently if a account fails to save it just displays an error with a
generic error message. This makes two changes to improve error messages
when connecting an account (currently only GitHub).

1. Display any and all error messages from ActiveRecord validations in
the flash alert.
2. Use a custom validator instead of the standard uniqueness validation
in order to add a custom error message that includes the account
provider and username in question so the user understands that this
account is already linked in Supermarket.