Cleanup styles from big refactor
I noticed a few minor things that needed to be cleaned up from the big
sidebar style clean up of 2014. Namely this changes so the gray
underlined links in the sidebar are only default within p tags and ul,
ol tags that have the pretty class, this prevents the cookbook listings
on the dashboard and such from taking on this style.