Admins are notified of requests to join their org
Users may only request to join an organization if they are not already
in the organization and if they have not already requested to join the
organization. The actual sending of the notifications happens in the
background because the number of admins an organization has is not
constrained.

The feature spec is marked as pending because the whole feature is not
complete. Once admins can accept (and decline) requests, the feature
will no longer be marked as pending.