Use the ICLA Signature record for names in table.
A user may or may not have a first and last name in their user record. The ICLA
Signature record for that user should _always_ have a first and last name (as
it is required). This updates the ICLA signers table to use the first and last
name from the signature record instead of the user record.