Set cookbook updated at date
We used to set some cookbook attributes in Cookbook#publish_version but
no longer do. We left save! in to bump Cookbook#updated_at when a new
version was published but it turns out this doesn't happen if no
attributes are changed so this explicitly sets the updated at date to
the current time.