Curry comments when PR authors become authorized
Previously, Curry would only add a label to the PR. Now, to ensure that
everyone who's subscribed to the PR's notifications knows that the PR
authors have been authorized, Curry also leaves a comment. Curry _only_
leaves a comment when the PR transitions to an authorized state; if the
PR is opened by authorized contributors, then Curry only adds the label
(no news is good news).