Adjust download cookbook button URL
Since this button is used on the cookbook version and cookbook view
rather than linking to the cookbook download path just link to the
download action for the current controller, either cookbooks or cookbook
versions. The URL will then adjust depending on if you're viewing
a cookbook or a cookbook version.