Implement irc_log chat channel redirect.
If the date requested is before August 8th, 2013, the user gets redirected to
a GitHub repo containing a dump of the chat logs. August 8th, 2013 is the first
known day for botbot.me to start having logs for freenode/chef.

If the date is August 8th, 2013 or after, the user gets redirected to the
requested date in the botbot.me logs.

If no date is specified, the user gets redirects to the botbot.me channel.

Things worth noting:

- This implementation assumes freenode as the server. The previous community
  site's controller implementation of logs did not reveal how servers were
  handled.
- If someone requests a channel that does not exist in botbot.me's logs, they
  will get redirected to botbot.me's 404. It feels like this redirect solution
  is best effort to get the user to where they want to go if the channel and
  date work. Since the URLs are not exposed within Supermarket's UI, I do not
  think we need to whitelist the channels or anything like that.