Display message about oc-id when users sign out.
Signing out of Supermarket does not sign users out of oc-id, which is the
expected behavior. However, it is not clear when signing out that this is the
behavior. This commit displays a message when a user is signed out letting them
know they need to go to id.opscode.com to signout completely.