Curry removes former commit authors from PRs
Until now, Curry would only ever add commit authors to its internal
record of each Pull Request. This results in incorrect records when
commit authors disappear from a Pull Request (for example, when commits
from several authors are sqaushed into one). Curry now compares the
post-import set of commit authors to the pre-import set and removes
commit authors who are in the latter but not in the former.