Create the public User show API.
A user and their related cookbooks can be viewed via the API by issuing a GET
request to /api/v1/users/:username.

Example:
  GET /api/v1/users/opscode

This returns the user, quite a bit of their data on their record and the
cookbooks they are associated with (owns, collaborates, follows). This does not
include Tools because there is no API available for them at this time.