POST to Fieri after a cookbook upload.
If Fieri is enabled and the FIERI_URL is configured, send a POST request to the
configured FIERI_URL after a cookbook version is successfully shared. This sends
along the cookbook name, the version and a link to the artifact in S3. Fieri
will then do what it needs to do and then send a POST request back to
Supermarket when it is done evaluating the cookbook.

This does not implement the end-point for Fieri to use, it just the first part
of making this all happen.