Create endpoint for receiving foodcritic results.
This implements the endpoint to receive the results from Fieri (the cookbook
evaluation service) and storing those results on the specified Cookbook Version.

Fieri should issue a POST request to Supermarket with the cookbook_name,
cookbook_version, foodcritic_failure, and foodcritic_feedback. The last two
get stored on the record.

This also requires specific params for evaluation endpoint. If the required
params are not present, an error will get raised. The controller rescues from
that error and renders a 400 error with why it failed.