Create the Tools show API endpoint
This allows access to the information for a specific Tool via the API.

An example request:

     GET /api/v1/tools/berkshelf

This will return the information for the berkshelf tool on Supermarket.

Unlike the rest of the API, this commit commit does not add view level specs. I
feel that the API spec properly tests the signature of what is output via the
API, which, to me, feels like the same expectations as the view specs.