Cookbook notification emails should be about the updated version.
Currently when a cookbook is updated a notification email goes out with
info for the latest cookbook version. This works in most cases but there
is a chance that the email will not actually pertain to the version that
triggered it. This changes the mailer so it accepts a cookbook version,
allowing each email to pertain to the exact version that was updated.

Closes #847.