Add CclaSig@latest_by_user to see multiple signees from an organization
Up to now, CclaSignatures have only ever been collected by unquing queries
on the organization. When we want to thank CCLA community members for
contributing, using the latest query uniqued on organization would mean
that multiple signees from a single organization would not be thanked.

If we want to get a list of all _users_ whose signatures are CCLA, not
ICLA, we need to DISTINCT on user_id, not organization_id. The difference
being do we want to display a list of organizations that have signed or
the people within them.