Add logrotate to the installer and a rule for nginx logs
nginx is configured to log to the standard access.log and error.log.
Convincing nginx to log to console in such a way to make runit's svlogd
able to manage it got more complicated than just adding logrotate to
noodle bowl. Precent for logrotate and nginx set in the Chef server, too.

* adds logrotate to the omninbus project to keep it separate from whatever
  logrotate might come on the host
* adds a log_management recipe to configure logrotate and schedule the
  custom supermarket logrotate to run via cron.hourly
* adds a logrotate rule to the nginx cookbook to lay down management
  of nginx's logs
* tests!