Fix sending USR1 signal to nginx on log rotations
Because we're using runit for managing omnibus services, there is no pid
file[1] for the nginx command to use to lookup to which process the
`reopen` command should be sent. This change is to have the postrotate
command call runit's `sv` to send USR1 to the nginx service[2] defined in
the service directory within the omnibus install.

[1] http://smarden.org/runit/benefits.html#supervision
[2] http://smarden.org/runit/sv.8.html under section "Commands"