Add detail from exception thrown during version validation
The default message was vague for an invalid dependency or platform version
constraint. For example:

```
Stove::Cli | >>>> The Chef Server did not understand the request because it was malformed.
{"error_code"=>"INVALID_DATA", "error_messages"=>["Version constraint is not a valid Chef version constraint"]}
```

The error above did not contain enough information for a cookbook author to
determine which of the versions they had declared in metadata was invalid.

This change includes the exception class and detail from the original
exception thrown about what is invalid about the data given. With this
change, the error above would appear as:

```
Stove::Cli | >>>> The Chef Server did not understand the request because it was malformed.
{"error_code"=>"INVALID_DATA", "error_messages"=>["Version constraint is not a valid Chef version constraint. Chef::Exceptions::InvalidCookbookVersion: '2' does not match 'x.y.z' or 'x.y'"]}
```

Updated spec expectations to use blocks to call out the action under test
and to check for a regex match among the validation errors instead of an
exact string inclusion.